+++ src/main/java/com/bcopstein/ex4_lancheriaddd_v1/Dominio/Entidades/ItemPedido.java
@@ -2,25 +2,24 @@ package com.bcopstein.ex4_lancheriaddd_v1.Dominio.Entidades;
 
 public class ItemPedido {
     private Produto item;
-    private int quantidade;
+    private long quantidade;
     private long id;
 
-    public ItemPedido( long id, Produto item, int quantidade) {
+    public ItemPedido(long id, Produto item, long quantidade) {
         this.id = id;
         this.item = item;
         this.quantidade = quantidade;
     }
 
-    
     public long getId() {
         return id;
     }
-    
+
     public Produto getItem() {
         return item;
     }
 
-    public int getQuantidade() {
+    public long getQuantidade() {
         return quantidade;
     }
 }
